requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Data manipulation and analysis
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration management
pyyaml>=6.0

# Statistical analysis
scikit-learn>=1.0.0

# Progress tracking (optional but useful)
tqdm>=4.62.0

# Testing (for development)
pytest>=6.2.0
pytest-cov>=2.12.0

# Type hints support
typing-extensions>=4.0.0

# For Jupyter notebooks (optional)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Documentation (optional)
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0 